requirements.txt
astroid==2.2.5
backports.csv==1.0.7
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-ckeditor==5.7.0
django-filter==2.1.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-material==1.5.6
django-material-admin==0.0.7
django-materialize==1.0.1a2
django-materializecss-form==1.1.14
djangorestframework==3.9.4
et-xmlfile==1.0.1
isort==4.3.17
jdcal==1.4.1
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.3.4
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
